astrbt_plugin_tianqi/requirements.txt
# 天气插件依赖项
aiohttp>=3.8.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
hypothesis>=6.0.0